+++ SAGA/database/__init__.py
@@ -4,4 +4,4 @@ SAGA.database
 This subpackage contains database-related routines, including Database
 """
 
-from .database import *
+from .core import *
