+++ SAGA/utils/__init__.py
@@ -4,4 +4,4 @@ SAGA.utils
 This subpackage collects some handy functions.
 """
 
-from .utils import *
+from .funcs import *
